# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserEntity" }
              }
            }
          }
        },
        "tags": ["Users"]
      },
      "patch": {
        "operationId": "UsersController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserEntity" }
              }
            }
          }
        },
        "tags": ["Users"]
      }
    },
    "/users/me/github/org": {
      "get": {
        "operationId": "UsersController_getGithubOrgs",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Users"]
      }
    },
    "/users/me/github/repo": {
      "get": {
        "operationId": "UsersController_getGithubProjects",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "org",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "summary": "Get All Github Repo",
        "tags": ["Users"]
      }
    },
...
    "/projects/{projectid}/tasks/me": {
      "get": {
        "operationId": "TasksController_listTasks",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "statusId",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "phaseId",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPaginationEntity"
                }
              }
            }
          }
        },
        "tags": ["Tasks"]
      },
      "post": {
        "operationId": "TasksController_createTask",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskFullEntity" }
              }
            }
          }
        },
        "tags": ["Tasks"]
      }
    },
    "/projects/{projectid}/tasks/me/{id}": {
      "patch": {
        "operationId": "TasksController_updateTask",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskFullEntity" }
              }
            }
          }
        },
        "tags": ["Tasks"]
      },
      "get": {
        "operationId": "TasksController_getMyTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TaskFullEntity" }
              }
            }
          }
        },
        "tags": ["Tasks"]
      },
      "delete": {
        "operationId": "TasksController_deleteReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Tasks"]
      }
    },
    "/projects/{projectid}/tasks/github/sync": {
      "post": {
        "operationId": "TasksController_createGithubIssue",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "summary": "Create Github Issue",
        "tags": ["Tasks"]
      }
    },
    "/projects/{projectid}/reports/me": {
      "get": {
        "operationId": "ReportsController_listReports",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["INFO", "LOW", "MEDIUM", "HIGH"],
              "type": "string"
            }
          },
          {
            "name": "issueType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "UI",
                "FUNCTIONAL",
                "PERFORMANCE",
                "SECURITY",
                "NETWORK",
                "DATA",
                "OTHER"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "INIT",
                "CONFIRMING",
                "CONFIRMED",
                "IN_PROCESSING",
                "REJECTED",
                "DONE",
                "REOPEN"
              ],
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "phaseId",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportPaginationEntity"
                }
              }
            }
          }
        },
        "tags": ["Reports"]
      },
      "post": {
        "operationId": "ReportsController_createReport",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateReportDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReportFullEntity" }
              }
            }
          }
        },
        "tags": ["Reports"]
      }
    },
    "/projects/{projectid}/reports/me/{id}": {
      "patch": {
        "operationId": "ReportsController_updateReport",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateReportDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReportFullEntity" }
              }
            }
          }
        },
        "tags": ["Reports"]
      },
      "get": {
        "operationId": "ReportsController_getMyReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ReportFullEntity" }
              }
            }
          }
        },
        "tags": ["Reports"]
      },
      "delete": {
        "operationId": "ReportsController_deleteReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Reports"]
      }
    },
    "/projects/{projectid}/reports/me/{id}/merge": {
      "post": {
        "operationId": "ReportsController_mergeReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MergeReportDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Reports"]
      }
    },
...
  },
  "info": {
    "title": "THESIS",
    "description": "The API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "RoleEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "category": {
            "type": "string",
            "enum": ["MEMBER", "GUEST", "OWNER"]
          },
          "name": { "type": "string" },
          "projectId": { "type": "number" }
        },
        "required": ["id", "category", "name", "projectId"]
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "email": { "type": "string" },
          "fullname": { "type": "string" },
          "avatar": { "type": "string" },
          "address": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "role": { "$ref": "#/components/schemas/RoleEntity" },
          "phoneNumber": { "type": "string" },
          "username": { "type": "string" },
          "githubId": { "type": "string" },
          "githubUsername": { "type": "string" }
        },
        "required": [
          "id",
          "email",
          "fullname",
          "createdAt",
          "phoneNumber",
          "username"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "phoneNumber": { "type": "string" },
          "address": { "type": "string" },
          "fullname": { "type": "string" },
          "avatar": { "type": "string" }
        },
        "required": ["username", "email", "phoneNumber", "fullname"]
      },
...
      "ProjectMemberEntity": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/UserCompactEntity" }
        },
        "required": ["user"]
      },
...
      "TaskListItemEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "createdBy": { "type": "number" },
          "estimateTime": { "type": "number" },
          "deadline": { "format": "date-time", "type": "string" },
          "priority": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "IMMEDIATE"]
          },
          "isPublic": { "type": "boolean" },
          "statusId": { "type": "number" },
          "categoryId": { "type": "number" },
          "references": { "type": "array", "items": { "type": "string" } },
          "assignedTo": { "type": "number" },
          "Assignee": { "$ref": "#/components/schemas/ProjectMemberEntity" },
          "ProjectMember": {
            "$ref": "#/components/schemas/ProjectMemberEntity"
          },
          "tags": { "type": "array", "items": { "type": "number" } },
          "Report": { "type": "object" },
          "reportId": { "type": "number" },
          "phaseId": { "type": "number" },
          "taskType": { "type": "string", "enum": ["GITHUB", "DEFAULT"] },
          "IssueGithub": { "$ref": "#/components/schemas/IssueGithubEntity" }
        },
        "required": [
          "id",
          "name",
          "description",
          "createdBy",
          "ProjectMember",
          "Report",
          "reportId"
        ]
      },
      "TaskPaginationEntity": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TaskListItemEntity" }
          }
        },
        "required": ["total", "items"]
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "estimateTime": { "type": "number" },
          "deadline": { "format": "date-time", "type": "string" },
          "priority": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "IMMEDIATE"]
          },
          "isPublic": { "type": "boolean" },
          "statusId": { "type": "number" },
          "categoryId": { "type": "number" },
          "references": { "type": "array", "items": { "type": "string" } },
          "assignedTo": { "type": "number" },
          "tags": { "type": "array", "items": { "type": "number" } },
          "attachments": { "type": "array", "items": { "type": "number" } },
          "reportId": { "type": "number" },
          "phaseId": { "type": "number" },
          "type": { "type": "string", "enum": ["GITHUB", "DEFAULT"] },
          "repoId": { "type": "number" }
        },
        "required": ["name", "description"]
      },
...
      "TaskFullEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "createdBy": { "type": "number" },
          "estimateTime": { "type": "number" },
          "deadline": { "format": "date-time", "type": "string" },
          "priority": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "IMMEDIATE"]
          },
          "isPublic": { "type": "boolean" },
          "statusId": { "type": "number" },
          "categoryId": { "type": "number" },
          "references": { "type": "array", "items": { "type": "string" } },
          "assignedTo": { "type": "number" },
          "Assignee": { "$ref": "#/components/schemas/ProjectMemberEntity" },
          "ProjectMember": {
            "$ref": "#/components/schemas/ProjectMemberEntity"
          },
          "tags": { "type": "array", "items": { "type": "number" } },
          "Report": { "type": "object" },
          "reportId": { "type": "number" },
          "phaseId": { "type": "number" },
          "taskType": { "type": "string", "enum": ["GITHUB", "DEFAULT"] },
          "IssueGithub": { "$ref": "#/components/schemas/IssueGithubEntity" },
          "TaskAttachment": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TaskAttachmentEntity" }
          },
          "TaskComment": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TaskCommentsEntity" }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "createdBy",
          "ProjectMember",
          "Report",
          "reportId",
          "TaskAttachment",
          "TaskComment"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "estimateTime": { "type": "number" },
          "deadline": { "format": "date-time", "type": "string" },
          "priority": {
            "type": "string",
            "enum": ["LOW", "MEDIUM", "HIGH", "IMMEDIATE"]
          },
          "isPublic": { "type": "boolean" },
          "statusId": { "type": "number" },
          "categoryId": { "type": "number" },
          "references": { "type": "array", "items": { "type": "string" } },
          "assignedTo": { "type": "number" },
          "tags": { "type": "array", "items": { "type": "number" } },
          "newAttachments": { "type": "array", "items": { "type": "array" } },
          "deleteAttachments": {
            "type": "array",
            "items": { "type": "array" }
          },
          "attachments": { "type": "array", "items": { "type": "number" } },
          "reportId": { "type": "number" },
          "phaseId": { "type": "number" }
        }
      },
...
      "ReportPaginationEntity": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReportListItemEntity" }
          }
        },
        "required": ["total", "items"]
      },
...
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string" },
          "severity": { "type": "string" },
          "isPublic": { "type": "boolean" },
          "assignedTo": { "type": "number" },
          "description": { "type": "string" },
          "stepsToReproduce": { "type": "string" },
          "expectedBehavior": { "type": "string" },
          "actualResult": { "type": "string" },
          "issueType": { "type": "string" },
          "additionInfo": { "type": "object" },
          "url": { "type": "string" },
          "status": { "type": "string" },
          "images": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CreateReportImageDto" }
          },
          "tags": { "type": "array", "items": { "type": "string" } },
          "phaseId": { "type": "number" }
        },
        "required": [
          "name",
          "severity",
          "description",
          "stepsToReproduce",
          "expectedBehavior",
          "actualResult",
          "issueType",
          "url"
        ]
      },
...
      "ReportFullEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["BUG", "FEEDBACK", "WISH"] },
          "description": { "type": "string" },
          "stepsToReproduce": { "type": "string" },
          "expectedBehavior": { "type": "string" },
          "actualResult": { "type": "string" },
          "issueType": {
            "type": "string",
            "enum": [
              "UI",
              "FUNCTIONAL",
              "PERFORMANCE",
              "SECURITY",
              "NETWORK",
              "DATA",
              "OTHER"
            ]
          },
          "severity": {
            "type": "string",
            "enum": ["INFO", "LOW", "MEDIUM", "HIGH"]
          },
          "assignedTo": { "type": "number" },
          "createdById": { "type": "number" },
          "projectId": { "type": "number" },
          "status": {
            "type": "string",
            "enum": [
              "INIT",
              "CONFIRMING",
              "CONFIRMED",
              "IN_PROCESSING",
              "REJECTED",
              "DONE",
              "REOPEN"
            ]
          },
          "url": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "assignee": { "$ref": "#/components/schemas/UserCompactEntity" },
          "createdBy": { "$ref": "#/components/schemas/UserCompactEntity" },
          "phaseId": { "type": "number" },
          "isProcessing": { "type": "boolean" },
          "isClosable": { "type": "boolean" },
          "ReportImage": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReportImageEntity" }
          },
          "ReportComment": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReportCommentsEntity" }
          },
          "DuplicateGroup": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReportDuplicateEntity" }
          },
          "children": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReportCompactEntity" }
          },
          "Task": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TaskCompactEntity" }
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "description",
          "createdById",
          "projectId",
          "status",
          "url",
          "createdAt",
          "updatedAt",
          "createdBy",
          "isClosable",
          "ReportImage",
          "ReportComment"
        ]
      },
      "UpdateReportDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string" },
          "severity": { "type": "string" },
          "isPublic": { "type": "boolean" },
          "assignedTo": { "type": "number" },
          "description": { "type": "string" },
          "stepsToReproduce": { "type": "string" },
          "expectedBehavior": { "type": "string" },
          "actualResult": { "type": "string" },
          "issueType": { "type": "string" },
          "additionInfo": { "type": "object" },
          "url": { "type": "string" },
          "status": { "type": "string" },
          "images": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CreateReportImageDto" }
          },
          "tags": { "type": "array", "items": { "type": "string" } },
          "phaseId": { "type": "number" },
          "newImages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CreateReportImageDto" }
          },
          "deleteImages": { "type": "array", "items": { "type": "number" } }
        }
      },
      "MergeReportDto": {
        "type": "object",
        "properties": {
          "childrenId": { "type": "number" },
          "type": { "type": "string" }
        },
        "required": ["childrenId", "type"]
      },
...
    }
  },
  "security": [{ "bearer": [] }]
}